requirements.txt
# Core dependencies
numpy>=1.19.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.5.0  # For model conversion
onnx>=1.10.0
onnxruntime>=1.8.0

# NeRF specific
pytorch3d>=0.6.0
torchsparse>=1.4.0
open3d>=0.13.0
trimesh>=3.9.0

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.0
pillow>=8.2.0
pyyaml>=6.0
tensorboard>=2.5.0
pytest>=6.2.5
jupyterlab>=3.0.0
scikit-image>=0.18.0
scikit-learn>=0.24.0

# Edge deployment
pycuda>=2021.1  # For NVIDIA Jetson devices
